[RF] Use unsigned int for index.

+++ fifo.h
@@ -1,15 +1,13 @@
 #ifndef FIFO_H
 #define FIFO_H
 
-#include <stdint.h>
-
 struct _fifo_node_t;
 
 typedef struct {
   size_t S;
   size_t W;
-  int64_t P __attribute__((aligned(64)));
-  int64_t C __attribute__((aligned(64)));
+  size_t P __attribute__((aligned(64)));
+  size_t C __attribute__((aligned(64)));
   struct _fifo_node_t * T __attribute__((aligned(64)));
 } fifo_t;
